景点详情点击进入第三方地图

+++ PeachTravel/SpotPoi.h
@@ -25,8 +25,8 @@
 @property (nonatomic, copy) NSString *kengdieUrl;
 @property (nonatomic) BOOL isMyFavorite;
 
-@property (nonatomic) long lat;
-@property (nonatomic) long lng;
+@property (nonatomic) double lat;
+@property (nonatomic) double lng;
 
 - (id)initWithJson:(id)json;
